src/software/backend/requirements.txt
asgiref==3.3.4
astroid==2.5.6
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
coverage==5.5
Django==3.2.3
djangorestframework==3.12.4
django-filter>=2.4.0
django-allauth>=0.44.0
dj-rest-auth>=2.1.7
python-dotenv>=0.17.0
djangorestframework_simplejwt>=4.7.1
flake8==3.9.2
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.8.2
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
